fix: resolve CI failures for Hytte-i6e8
Move computePickerPosition and its types to pickerPosition.ts so
ReactionPicker.tsx only exports a component, satisfying the
react-refresh/only-export-components lint rule that CI enforces.
